backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
pillow>=10.0.0
numpy>=1.24.0
slowapi>=0.1.8
tensorflow>=2.14.0
opencv-python-headless>=4.8.0
